Feat: 뷰에서 외부 참조 필요 없는 상수 및 변수들 접근제어 및 소수 선언 형식 통일

+++ Solidner/Views/Onboarding/NickNameView.swift
@@ -8,9 +8,9 @@
 import SwiftUI
 
 struct NickNameView: View {
-    private let textFieldTopPadding = CGFloat(16)
-    private let warningMessageTopPadding = CGFloat(12)
-    private let warningMessageLeadingPadding = CGFloat(4)
+    private let textFieldTopPadding = 16.0
+    private let warningMessageTopPadding = 12.0
+    private let warningMessageLeadingPadding = 4.0
     private let warningMessageFontSize = 11.5
     private let buttonUpDuration = 0.1
     var nickNameViewCase = NickNameViewCase.babyName
